slight improvement for patrykel's changes
