+++ bin/www
@@ -4,8 +4,10 @@
 let debug = require('debug')('hero-dashboard:server');
 let server;
 
-let heroku = require('heroku-client');
-let app = require('./../src/app')();
+let Heroku = require('heroku-client');
+let herokuRequests = require('./../src/heroku-requests')(new Heroku({token: process.env.HEROKU_API_TOKEN}));
+let cache = require('./../src/cache/in-memory-cache')();
+let app = require('./../src/app')(herokuRequests, cache);
 let http = require('http');
 
 let port = normalizePort(process.env.PORT || '3000');
