+++ test/stubs/heroku-requests-stub.js
@@ -1,0 +1,7 @@
+module.exports = () => {
+    return {
+        getAllApps: () => {
+            return require('./../resources/app1-data.json');
+        }
+    };
+};
